refactor(Export): Move createStudentWorkExportRow function
