+++ src/assets/wise5/classroomMonitor/dataExport/strategies/AbstractDataExportStrategy.ts
@@ -7,6 +7,7 @@ import { TeacherProjectService } from '../../../services/teacherProjectService';
 import { DataExportComponent } from '../data-export/data-export.component';
 import { DataExportContext } from '../DataExportContext';
 import { DataExportStrategy } from './DataExportStrategy';
+import { removeHTMLTags } from '../../../common/string/string';
 
 export abstract class AbstractDataExportStrategy implements DataExportStrategy {
   context: DataExportContext;
@@ -119,6 +120,33 @@ export abstract class AbstractDataExportStrategy implements DataExportStrategy {
 
   insertColumnAtEnd(headerRow: string[], columnName: string): void {
     headerRow.push(columnName);
+  }
+
+  /**
+   * Get the plain text representation of the student work.
+   * @param componentState {object} A component state that contains the student work.
+   * @returns {string} A string that can be placed in a csv cell.
+   */
+  getStudentDataString(componentState: any): string {
+    /*
+     * In Excel, if there is a cell with a long string and the cell to the
+     * right of it is empty, the long string will overlap onto cells to the
+     * right until the string ends or hits a cell that contains a value.
+     * To prevent this from occurring, we will default empty cell values to
+     * a string with a space in it. This way all values of cells are limited
+     * to displaying only in its own cell.
+     */
+    let studentDataString = ' ';
+    let componentType = componentState.componentType;
+    let componentService = this.controller.componentServiceLookupService.getService(componentType);
+    if (componentService != null && componentService.getStudentDataString != null) {
+      studentDataString = componentService.getStudentDataString(componentState);
+      studentDataString = removeHTMLTags(studentDataString);
+      studentDataString = studentDataString.replace(/"/g, '""');
+    } else {
+      studentDataString = componentState.studentData;
+    }
+    return studentDataString;
   }
 
   generateExportFileName(
